+++ src/WindowsMcp/Program.cs
@@ -75,6 +75,7 @@ internal static class Program
         builder.Services.AddSingleton<IDiskService, DiskService>();
         builder.Services.AddSingleton<ISecurityService, SecurityService>();
         builder.Services.AddSingleton<IFirewallService, FirewallService>();
+        builder.Services.AddSingleton<ICertStoreService, CertStoreService>();
         builder.Services.AddSingleton<IEnvService, EnvService>();
         builder.Services.AddSingleton<IPowerService, PowerService>();
         builder.Services.AddSingleton<INotificationService, NotificationService>();